Fix bug and tidy up timeGroupMore
* Fixed bug with switches in timeGroups applying too soon, while the
  switch button disappeared
* Renamed T_TimeDataUnit properties to add information about timing:
  two properties reflect conditions during production, two are about
  the end result of purchasing the upgrades and one is speculative, so
  clarifying timing should help prevent mix-ups in future
* Refactored timeGroupMore to break up into multiple files
* Moved timeGroupMore into its own subfolder
* Continued efforts to consistently reserve "get" for outside sources
  of data
* Refactored deepCopy to make use of structuredClone(), since I just
  learned that exists
