+++ app/components/timeGroupMore/subcomponents/subheading.tsx
@@ -1,0 +1,6 @@
+export default function TimeGroupMoreSubheading({text} 
+    : {text : string})
+    : JSX.Element {
+
+    return <h4 className={"text-xs self-start font-bold"}>{text}</h4>;
+}
